Create tic_tac_toe_game.c
Tic-Tac-Toe game can be 
played by two players where the square block (3 x 3) can 
be filled with a cross (X) or a circle (O). The game will 
toggle between the players by giving the chance for each 
player to mark their move. When one of the players make 
a combination of 3 same markers in a horizontal, vertical 
or diagonal line the program will display which player has 
won, whether X or O. In this program, we implement a 3x3 
tic-tac-toe game. The game is designed so that two players 
can play tic-tac-toe using C-language. The program will 
contain a display function and a select function to place 
the symbol as well as toggle between the symbols allowing 
each player a turn to play the game. The program will 
update after each player makes their move and check for 
the conditions of game as it goes on. Overall program 
works without any bugs and is able to use